Clips and drawings fixes,
fixed bad tab on show settings before conversion
fixed init Direct X crash
Disabled scaling dpi crashing code

+++ Kaiplayer/Visuals.h
@@ -351,7 +351,7 @@ public:
 	int FindPoint(int pos, wxString type, bool nextStart = false, bool fromEnd = false);
 	ClipPoint FindSnapPoint(const ClipPoint &pos, size_t pointToSkip/*, bool coeff = false*/);
 	void OnKeyPress(wxKeyEvent &evt);
-	void OnMoveSelected(int x, int y);
+	void OnMoveSelected(float x, float y);
 	int CheckCurve(int pos, bool checkSpline = true);
 	void AppendClipMask(wxString *mask);
 	void CreateClipMask(wxString *clip, wxString *clipTag = NULL);
